# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Connect | Server",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/company-groups/{companyGroupId}": {
      "put": {
        "operationId": "updateCompanyGroup",
        "summary": "Update company group",
        "tags": [
          "Company Group"
        ],
        "description": "Update company group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "document": {
                    "description": "Brazilian CNPJ",
                    "type": "string",
                    "minLength": 14,
                    "maxLength": 14
                  }
                },
                "required": [
                  "name",
                  "document"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyGroupId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCompanyGroup",
        "summary": "Delete company group",
        "tags": [
          "Company Group"
        ],
        "description": "Delete company group",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyGroupId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getCompanyGroup",
        "summary": "Get company group",
        "tags": [
          "Company Group"
        ],
        "description": "Get company group",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyGroupId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "document": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deletedAt": {
                          "nullable": true,
                          "type": "string",
                          "format": "date-time"
                        },
                        "companiesCount": {
                          "type": "number"
                        },
                        "phones": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "cuid",
                                "pattern": "^[cC][^\\s-]{8,}$"
                              },
                              "number": {
                                "type": "string"
                              },
                              "name": {
                                "nullable": true,
                                "type": "string"
                              },
                              "isWhatsapp": {
                                "type": "boolean"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "id",
                              "number",
                              "name",
                              "isWhatsapp",
                              "createdAt"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "document",
                        "createdAt",
                        "updatedAt",
                        "deletedAt",
                        "companiesCount",
                        "phones"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/units/{unitId}": {
      "put": {
        "operationId": "updateUnit",
        "summary": "Update unit",
        "tags": [
          "Unit"
        ],
        "description": "Update unit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteUnit",
        "summary": "Delete unit",
        "tags": [
          "Unit"
        ],
        "description": "Delete unit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getUnit",
        "summary": "Get unit",
        "tags": [
          "Unit"
        ],
        "description": "Get unit",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "basesCount": {
                          "type": "number"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "company": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "cuid",
                              "pattern": "^[cC][^\\s-]{8,}$"
                            },
                            "name": {
                              "type": "string"
                            },
                            "document": {
                              "type": "string"
                            },
                            "companyGroup": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "format": "cuid",
                                  "pattern": "^[cC][^\\s-]{8,}$"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "document": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "document"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "document",
                            "companyGroup"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "basesCount",
                        "createdAt",
                        "updatedAt",
                        "company"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/units/{unitId}/phones/{phoneId}": {
      "delete": {
        "operationId": "deleteUnitPhoneNumber",
        "summary": "Delete unit phone number",
        "tags": [
          "Unit"
        ],
        "description": "Delete unit phone number",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "unitId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "phoneId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}